extracted scroll stalker component from TableOfContents example
